+++ taskmonitor/tests/core/test_cached_reports.py
@@ -183,7 +183,7 @@ class TestExceptionsOverTime(TestCase):
     def setUp(self) -> None:
         cache.clear()
 
-    def test_should_create_queue_report(self):
+    def test_should_create_report(self):
         # given
         start_dt = dt.datetime(2023, 1, 1, 12, 0, tzinfo=utc)
         TaskLogFactory(
@@ -207,3 +207,34 @@ class TestExceptionsOverTime(TestCase):
         # then
         self.assertEqual(result[0]["name"], "Alpha")
         self.assertEqual(result[1]["name"], "Bravo")
+
+
+class TestAppFailuresOverTime(TestCase):
+    def setUp(self) -> None:
+        cache.clear()
+
+    def test_should_create_report(self):
+        # given
+        start_dt = dt.datetime(2023, 1, 1, 12, 0, tzinfo=utc)
+        TaskLogFactory(
+            received=start_dt,
+            started=start_dt,
+            timestamp=start_dt + dt.timedelta(seconds=5),
+            current_queue_length=30,
+            exception="Bravo",
+            state=TaskLog.State.FAILURE,
+        )
+        TaskLogFactory(
+            received=start_dt,
+            started=start_dt,
+            timestamp=start_dt + dt.timedelta(seconds=5),
+            current_queue_length=30,
+            exception="Alpha",
+            state=TaskLog.State.FAILURE,
+        )
+        start_dt += dt.timedelta(minutes=5)
+        report = cached_reports.AppFailuresOverTime()
+        # when
+        result = report._calc_data()
+        # then
+        self.assertTrue(result)